Fixed Random Crash Issue
Found a solution online that is linked below.  Basically, the program
would read the position of the element that it is supposed to click
before the page would finish loading.  The element would then move as
other things on the page finish loading.  Then when the program would go
to click the link, it was no longer in the same spot.

To fix this I added a one second delay before the 'click link' command
to allow the page time to finish loading.

https://code.google.com/p/selenium/issues/detail?id=2766